refactor: 상속을 통해 관심사 분리
