+++ toby-spring/src/main/java/com/example/tobyspring/PaymentService.java
@@ -1,16 +1,10 @@
 package com.example.tobyspring;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.math.BigDecimal;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.time.LocalDateTime;
-import java.util.stream.Collectors;
 
-public class PaymentService {
+public abstract class PaymentService {
 
     public Payment prepare(
             Long orderId, String currency, BigDecimal foreignCurrencyAmount
@@ -24,23 +18,6 @@ public class PaymentService {
 
     }
 
-    private BigDecimal getExchangeRate(String currency) throws IOException {
-        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
-        HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
-        BufferedReader br = new BufferedReader(new InputStreamReader(httpURLConnection.getInputStream()));
-        String response = br.lines().collect(Collectors.joining());
-        br.close();
-
-        ObjectMapper mapper = new ObjectMapper();
-        ExchangeRateVO exchangeRateVO = mapper.readValue(response, ExchangeRateVO.class);
-        BigDecimal exchangeRate = exchangeRateVO.rates().get("KRW");
-        return exchangeRate;
-    }
-
-    public static void main(String[] args) throws IOException {
-        PaymentService paymentService = new PaymentService();
-        Payment payment = paymentService.prepare(100L, "USD", BigDecimal.valueOf(50.7));
-        System.out.println(payment);
-    }
+    abstract BigDecimal getExchangeRate(String currency) throws IOException;
 
 }
