+++ toby-spring/src/main/java/com/example/tobyspring/SimpleExchangeRatePaymentService.java
@@ -1,0 +1,20 @@
+package com.example.tobyspring;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+
+public class SimpleExchangeRatePaymentService extends PaymentService {
+
+
+    @Override
+    BigDecimal getExchangeRate(String currency) throws IOException {
+
+        if (currency.equals("USD")) {
+            return BigDecimal.valueOf(1000);
+        }
+
+        throw new IllegalArgumentException("지원되지 않는 통화입니다.");
+
+    }
+
+}
