+++ toby-spring/src/main/java/com/example/tobyspring/WebApiExchangeRatePaymentService.java
@@ -1,0 +1,28 @@
+package com.example.tobyspring;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.math.BigDecimal;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.util.stream.Collectors;
+
+public class WebApiExchangeRatePaymentService extends PaymentService {
+
+    @Override
+    BigDecimal getExchangeRate(String currency) throws IOException {
+        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
+        HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
+        BufferedReader br = new BufferedReader(new InputStreamReader(httpURLConnection.getInputStream()));
+        String response = br.lines().collect(Collectors.joining());
+        br.close();
+
+        ObjectMapper mapper = new ObjectMapper();
+        ExchangeRateVO exchangeRateVO = mapper.readValue(response, ExchangeRateVO.class);
+        BigDecimal exchangeRate = exchangeRateVO.rates().get("KRW");
+        return exchangeRate;
+    }
+
+}
